refactor: extract bundle material adaptation
